testingdataflowimp-main/requirements.txt
pyarrow==0.15.1
numpy==1.20.0
six==1.13.0
apache-beam==2.30.0
google-cloud
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub
apitools
google-apitools
http.client
